Update GameClassesv2.pyde
updated Kirito's movements and changed the image fetching method, but it still needs a lot of work (I have an idea, but I'm still writing out the pseudo code for it atm). most likely, once the player class is finished, the rest of the game should be pretty simple once we add your background classes and set up the stages! I've looked into it, and I think we can use if statements to make the levels change (i.e. if player reaches a certain X, then the background updating scrolling stops and the player can go to the end of the screen and that's when the level changes and backgrounds change and new entities are added).