Update GameClassesv2.pyde
added some bits for the enemy, also added some of your programming for the keys. I haven't integrated your code completely, so it's still a work in progress. I found a bug in the hitbox, so I'll be trying to fix that soon. Have some fun ideas for the boss, will code soon as well!